% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fitmssv}
\alias{fitmssv}
\title{Fit a bivariate structural stochastic volatility model.}
\usage{
fitmssv(y, x, start = c(0.95, 0.95, 0.3, 0.3, 0.02))
}
\arguments{
\item{y}{a bivariate numeric or time series containing log returns. The first column
should contain the "structural" series, while the second corresponds to the market.}

\item{x}{an explanatory variable, presumably the log of a leverage multiplier.}

\item{start}{starting parameters for the optimization.}
}
\value{
A list containing the output from the solver ("model") and the outputs from 
the Kalman filter ("fit").
}
\description{
Fit a bivariate structural stochastic volatility model.
}
