% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fitssv}
\alias{fitssv}
\title{Fit a structural stochastic volatility model.}
\usage{
fitssv(y, x, N = 5, start = c(0.95, 0.3))
}
\arguments{
\item{y}{a numeric vector or time series containing log returns.}

\item{x}{an explanatory variable, presumably the log of a leverage multiplier.}

\item{N}{number of importance samples to draw for the Monte Carlo ll evaluation.}

\item{start}{starting parameters for the optimization.}
}
\value{
A list containing the output from the solver ("model") and the outputs from 
the Kalman filter and monte carlo evaluation routine ("fit").
}
\description{
Fit a structural stochastic volatility model.
}
