% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{levmulttest}
\alias{levmulttest}
\title{Test for misspecification in the form of an excluded leverage multiplier.}
\usage{
levmulttest(data, lmt, model)
}
\arguments{
\item{data}{The data to be tested.}

\item{model}{The null model.}
}
\value{
A test statistic with an asymptotic chi^2(1) distribution under the null model.
}
\description{
Test for misspecification in the form of an excluded leverage multiplier.
}
