% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{llevtest2}
\alias{llevtest2}
\title{Test for a leverage effect in the data based on the log garch representation.}
\usage{
llevtest2(data)
}
\arguments{
\item{data}{The data to be tested.}

\item{model}{The null model.}
}
\value{
A test statistic with an asymptotic chi^2(1) distribution under the null model.
}
\description{
Test for a leverage effect in the data based on the log garch representation.
}
