% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{llratiotest}
\alias{llratiotest}
\title{Likelihood ratio test for two competing stochastic volatility models.}
\usage{
llratiotest(model0, model1)
}
\arguments{
\item{model0}{The null model}

\item{model1}{The alternative model}
}
\value{
The likelihood ratio test statistic, here presumably asymptotically
distributed as chi^2(1) under the null.
}
\description{
Likelihood ratio test for two competing stochastic volatility models.
}
