% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{mssverrors}
\alias{mssverrors}
\title{MSSV Errors}
\usage{
mssverrors(model, y, x)
}
\arguments{
\item{model}{Output "model" from function "fitmssv".}

\item{y}{The bivariate time series of log returns that the model has been fitted to.}

\item{x}{The explanatory variable for the model of the first column in y, presumably 
a log leverage multiplier.}
}
\description{
Calculate numerical "sandwich" variance-covariance errors 
of a multivariate stochastic volatility model.
}
