% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{simdevent}
\alias{simdevent}
\title{simdevent}
\usage{
simdevent(
  N = 1000,
  len = 500,
  pars = c(-10, 0.95, 0.3, -12, 0.9, 0.2, 0.9, 1),
  Ain = 80,
  Ein = 60,
  K = 60,
  r = 0.001,
  uv = 5e-04,
  ttv = 5,
  thd = 0.8,
  plot = T,
  trim = T,
  trimquants = c(0.01, 0.99),
  crisisret = -0.1,
  seed = NULL
)
}
\arguments{
\item{N}{Integer, the number of time series to use for the simulation.}

\item{len}{Integer, the length of the simulated time period.}

\item{pars}{A vector with the parameter values of the structural multivariate stochastic volatility model.
Should be in the order: c(mu_1, beta_1, sigma_1, mu_2, beta_2, sigma_2, rho, phi),
where subscript 1 corresponds to the "structural" series that has an explanatory variable in the state equation.}

\item{Ain}{Numeric, initial asset value.}

\item{Ein}{Numeric, initial equity value.}

\item{K}{Numeric, debt value over the time period.}

\item{r}{Numeric, the risk-free rate of interest}

\item{uv}{Numeric, the daily unconditional variance of the assets over the period.}

\item{ttv}{Numeric, time to maturity of the debt over the period (measured in years).}

\item{thd}{Numeric, specifies a threshold leverage ratio that the user wants to monitor. Used to draw a vertical abline in the histogram.}

\item{plot}{Boolean, should a plot be created (T) or not (F)?}

\item{trim}{Boolean, should the simulated terminal leverage ratios be trimmed?}

\item{trimquants}{Vector indicating the quantiles to trim at if argument "trim" is set to TRUE.}
}
\value{
Plots the histogram with a vertical abline for a user-specified threshold and returns the simulated minimum leverage ratios.
}
\description{
Simulate the minimum leverage ratio during some period given inputs.
}
