% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{simintprob}
\alias{simintprob}
\title{simintprob}
\usage{
simintprob(
  N = 1000,
  len = 500,
  pars = c(-10, 0.95, 0.3, -12, 0.9, 0.2, 0.9, 1),
  Ain = 80,
  Ein = 60,
  K = 60,
  r = 0.001,
  uv = 5e-04,
  ttv = 5,
  lower = 0.4,
  upper = 0.6,
  seed = NULL,
  crisis = F,
  crisisret = -0.1
)
}
\arguments{
\item{len}{Integer, the length of the simulated time period.}

\item{pars}{A vector with the parameter values of the structural multivariate stochastic volatility model.}

\item{Ain}{Numeric, initial asset value.}

\item{Ein}{Numeric, initial equity value.}

\item{K}{Numeric, debt value over the time period.}

\item{r}{Numeric, the risk-free rate of interest}

\item{uv}{Numeric, the daily unconditional variance of the assets over the period.}

\item{ttv}{Numeric, time to maturity of the debt over the period (measured in years).}

\item{crisis}{Boolean, only keeps paths where the market return is below some threshold during the period.}

\item{crisisret}{The return that defines a crisis and sets Boolean crisis to TRUE.}
}
\value{
The probability that the leverage ratio lies in the specified range at the end of the period.
}
\description{
Simulate the probability that leverage ratio lies in some range at the end of the period.
}
