% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{simstrucsystem}
\alias{simstrucsystem}
\title{simstrucsystem}
\usage{
simstrucsystem(
  len = 2000,
  pars = c(-10, 0.95, 0.3, -12, 0.9, 0.2, 0.9),
  Ain = 100,
  Ein = 80,
  K = 20,
  r = 0.001,
  uv = 5e-04,
  ttv = 5
)
}
\description{
Simulate a bivariate structural volatility model.
}
