% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihoods.R
\name{sskloglik}
\alias{sskloglik}
\title{Calculate the Sandmann and Koopman log likelihood}
\usage{
sskloglik(param, yt, xt, N = 5, fit = T)
}
\description{
Calculate the Sandmann and Koopman log likelihood
}
