% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{simstrucsystem}
\alias{simstrucsystem}
\title{simstrucsystem}
\usage{
simstrucsystem(
  len = 2000,
  pars = c(-10, 0.95, 0.3, -12, 0.9, 0.2, 0.9, 1),
  Ain = 100,
  Ein = 80,
  K = 20,
  r = 0.001,
  uv = 5e-04,
  ttv = 5
)
}
\arguments{
\item{len}{Integer, the length of the simulated time period.}

\item{pars}{A vector with the parameter values of the structural multivariate stochastic volatility model.}

\item{Ain}{Numeric, initial asset value.}

\item{Ein}{Numeric, initial equity value.}

\item{K}{Numeric, debt value over the time period.}

\item{r}{Numeric, the risk-free rate of interest}

\item{uv}{Numeric, the daily unconditional variance of the assets over the period.}

\item{ttv}{Numeric, time to maturity of the debt over the period (measured in years).}
}
\value{
A list containing relevant simulated quantities.
}
\description{
Simulate a bivariate structural volatility model.
}
